+++ src/platform/linux/procfs_reader.hpp
@@ -18,7 +18,7 @@ public:
     std::optional<ProcessInfo> get_process_info(int pid, int64_t total_memory);
 
     static std::vector<ThreadInfo> get_threads(int pid);
-    static std::string get_thread_stack(int pid, int tid);
+    static std::string get_thread_stack(int pid, int64_t tid);
 
     static std::vector<FileHandleInfo> get_file_handles(int pid);
